fix: DCA每日定投整数截断导致只触发一次 + spendableCNY误拦截
- DaysSinceLastMacro int→float64: 1h K线间隔0.0417天在int除法中截断为0
- macro_engine: base DCA跳过spendableCNY clamp(新钱注入无需可用余额)
- step.go: MonthlyInject>0时总是运行macro引擎

+++ internal/quant/data.go
@@ -135,7 +135,7 @@ type MacroDecisionInput struct {
 	MonthlyInject        float64 // monthly injection amount (from SpawnPoint)
 	TimeDilation         float64 // from market state
 	PriceDeviation       float64 // (Price - EMA_long) / EMA_long
-	DaysSinceLastMacro   int     // days since last macro action
+	DaysSinceLastMacro   float64 // days since last macro action (fractional)
 	MacroIntervalDays    int     // base DCA interval days (from chromosome)
 	MacroAccelThreshold  float64 // price deviation trigger threshold
 	MacroAccelMultiplier float64 // acceleration multiplier
